[REF] Criação da classe ComunicacaoNFE com base na ComunicacaoSefaz
